# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Steward System",
        "description": "Documentação",
        "contact": {
            "email": "anaimayer3@gmail.com"
        },
        "license": {
            "name": "Proprietary"
        },
        "version": "0.1.0"
    },
    "paths": {
...
        "/user": {
            "get": {
                "tags": [
                    "Usuário"
                ],
                "summary": "Listar",
                "description": "Lista usuários cadastrados no BD. Necessário permissão admin.",
                "operationId": "52cf2b0b31b1d862f0ed91d8e1d00221",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "limit": {
                                        "type": "integer"
                                    },
                                    "offset": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "limit": 15,
                                    "offset": 0
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "usuarios": {
                                            "type": "json",
                                            "example": {
                                                "usuarios": {
                                                    "uuid": "7a02d1a0-1bd7-4d55-a32d-05e6ad3bee10",
                                                    "email": "exemplo@mail.com.br",
                                                    "email_verified_at": null,
                                                    "change_password": true,
                                                    "created_at": "2024-04-25T11:43:40.000000Z",
                                                    "updated_at": "2024-04-25T11:43:40.000000Z",
                                                    "admin": false
                                                }
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Usuário"
                ],
                "summary": "Cadastrar",
                "operationId": "2b65419508f430646443a589c50902be",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "perfil": {
                                        "type": "json"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "fulano@exemplo.com.br",
                                    "password": "S4af3P4assword",
                                    "perfil": {
                                        "nome": "Fulano da Silva",
                                        "cpf": 61141289075,
                                        "cpf_responsavel": false,
                                        "telefone": "(48) 9-9867-3412",
                                        "altura": 1.75,
                                        "avatar_url": "https://t3.ftcdn.net/jpg/05/53/79/60/360_F_553796090_XHrE6R9jwmBJUMo9HKl41hyHJ5gqt9oz.jpg"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user_uuid": {
                                            "type": "string",
                                            "example": "7a02d1a0-1bd7-4d55-a32d-05e6ad3bee10"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/{uuid}": {
            "get": {
                "tags": [
                    "Usuário"
                ],
                "summary": "Busca UUID",
                "description": "Busca um usuário com base no UUID passado na URL",
                "operationId": "722b0f0b8cc2880ea09ffb89634eab80",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "uuid": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "uuid": "7a02d1a0-1bd7-4d55-a32d-05e6ad3bee10"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "usuario": {
                                            "type": "json",
                                            "example": {
                                                "uuid": "7a02d1a0-1bd7-4d55-a32d-05e6ad3bee10",
                                                "email": "exemplo@mail.com.br",
                                                "email_verified_at": null,
                                                "change_password": true,
                                                "created_at": "2024-04-25T11:43:40.000000Z",
                                                "updated_at": "2024-04-25T11:43:40.000000Z",
                                                "admin": false,
                                                "perfil": {
                                                    "nome": "Usuário Exemplo",
                                                    "cpf": "61141289075",
                                                    "cpf_responsavel": false,
                                                    "telefone": "48998673412",
                                                    "altura": "1.75",
                                                    "avatar_url": "https://t3.ftcdn.net/jpg/05/53/79/60/360_F_553796090_XHrE6R9jwmBJUMo9HKl41hyHJ5gqt9oz.jpg"
                                                }
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Usuário"
                ],
                "summary": "Atualizar",
                "description": "Atualiza um usuário",
                "operationId": "7df442a9a24d0d830ddf6232b18b7a0a",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "uuid": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "uuid": "7a02d1a0-1bd7-4d55-a32d-05e6ad3bee10"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "No content"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Usuário"
                ],
                "summary": "Remover",
                "description": "Deleta um usuário (soft delete)",
                "operationId": "52821b050c6133e122f92860df70e90e",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "uuid": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "uuid": "7a02d1a0-1bd7-4d55-a32d-05e6ad3bee10"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "No content"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Usuário"
                ],
                "summary": "Admin",
                "description": "Atualiza o parâmetro admin de um usuário",
                "operationId": "48d1d621387ada480a1baf1d21081350",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "uuid": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "uuid": "7a02d1a0-1bd7-4d55-a32d-05e6ad3bee10"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "No content"
                    }
                }
            }
        }
    }
}